src/productcatalogservice/agent/requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-health-checking>=1.60.0

# REST server (for testing and comparison)
flask>=3.0.0

# Utilities
protobuf>=4.25.0